+++ src/Adapter/Canton/CantonWithPlaces.mjs
@@ -1,6 +1,0 @@
-/** @typedef {import("./Canton.mjs").Canton} Canton */
-/** @typedef {import("../Place/Place.mjs").Place} Place */
-
-/**
- * @typedef {Canton & {places: Place[]}} CantonWithPlaces
- */