+++ src/Adapter/FormTitle/FormTitleElement.mjs
@@ -1,5 +1,3 @@
-import { ELEMENT_TAG_NAME_PREFIX } from "../Element/ELEMENT_TAG_NAME_PREFIX.mjs";
-
 /** @typedef {import("../../Libs/flux-css-api/src/Adapter/Api/CssApi.mjs").CssApi} CssApi */
 
 const __dirname = import.meta.url.substring(0, import.meta.url.lastIndexOf("/"));
@@ -68,6 +66,6 @@ export class FormTitleElement extends HTMLElement {
     }
 }
 
-export const FORM_TITLE_ELEMENT_TAG_NAME = `${ELEMENT_TAG_NAME_PREFIX}form-title`;
+export const FORM_TITLE_ELEMENT_TAG_NAME = "flux-studis-selfservice-form-title";
 
 customElements.define(FORM_TITLE_ELEMENT_TAG_NAME, FormTitleElement);