+++ src/Adapter/Title/TitleElement.mjs
@@ -1,5 +1,3 @@
-import { ELEMENT_TAG_NAME_PREFIX } from "../Element/ELEMENT_TAG_NAME_PREFIX.mjs";
-
 /** @typedef {import("../../Libs/flux-css-api/src/Adapter/Api/CssApi.mjs").CssApi} CssApi */
 
 const __dirname = import.meta.url.substring(0, import.meta.url.lastIndexOf("/"));
@@ -60,6 +58,6 @@ export class TitleElement extends HTMLElement {
     }
 }
 
-export const TITLE_ELEMENT_TAG_NAME = `${ELEMENT_TAG_NAME_PREFIX}title`;
+export const TITLE_ELEMENT_TAG_NAME = "flux-studis-selfservice-title";
 
 customElements.define(TITLE_ELEMENT_TAG_NAME, TitleElement);