+++ src/Adapter/UniversityEntranceQualification/UniversityEntranceQualificationElement.mjs
@@ -1,21 +1,17 @@
-import { ELEMENT_TAG_NAME_PREFIX } from "../Element/ELEMENT_TAG_NAME_PREFIX.mjs";
 import { FormElement } from "../Form/FormElement.mjs";
 import { MandatoryElement } from "../Mandatory/MandatoryElement.mjs";
 import { PAGE_UNIVERSITY_ENTRANCE_QUALIFICATION } from "../Page/PAGE.mjs";
 import { TitleElement } from "../Title/TitleElement.mjs";
+import { UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_CERTIFICATE, UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_CERTIFICATE_CANTON, UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_CERTIFICATE_COUNTRY, UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_CERTIFICATE_PLACE, UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_CERTIFICATE_TYPE, UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_ISSUE_YEAR, UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_MATURA_CANTON, UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_UPPER_SECONDARY_SCHOOL } from "./UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE.mjs";
 
 /** @typedef {import("../Post/backFunction.mjs").backFunction} backFunction */
-/** @typedef {import("../Canton/CantonWithPlaces.mjs").CantonWithPlaces} CantonWithPlaces */
-/** @typedef {import("../Canton/CantonWithSchools.mjs").CantonWithSchools} CantonWithSchools */
-/** @typedef {import("../CertificateType/CertificateTypeWithCertificates.mjs").CertificateTypeWithCertificates} CertificateTypeWithCertificates */
-/** @typedef {import("../Certificate/CertificateWithCantons.mjs").CertificateWithCantons} CertificateWithCantons */
 /** @typedef {import("./chosenUniversityEntranceQualificationFunction.mjs").chosenUniversityEntranceQualificationFunction} chosenUniversityEntranceQualificationFunction */
-/** @typedef {import("../Country/CountryWithCantons.mjs").CountryWithCantons} CountryWithCantons */
 /** @typedef {import("../../Libs/flux-css-api/src/Adapter/Api/CssApi.mjs").CssApi} CssApi */
 /** @typedef {import("../../Service/Label/Port/LabelService.mjs").LabelService} LabelService */
 /** @typedef {import("../../Libs/flux-localization-api/src/Adapter/Api/LocalizationApi.mjs").LocalizationApi} LocalizationApi */
-/** @typedef {import("../School/SchoolWithCountries.mjs").SchoolWithCountries} SchoolWithCountries */
 /** @typedef {import("./UniversityEntranceQualification.mjs").UniversityEntranceQualification} UniversityEntranceQualification */
+/** @typedef {import("./UniversityEntranceQualificationData.mjs").UniversityEntranceQualificationData} UniversityEntranceQualificationData */
+/** @typedef {import("./UniversityEntranceQualificationSelectOption.mjs").UniversityEntranceQualificationSelectOption} UniversityEntranceQualificationSelectOption */
 
 const __dirname = import.meta.url.substring(0, import.meta.url.lastIndexOf("/"));
 
@@ -24,22 +20,6 @@ export class UniversityEntranceQualificationElement extends HTMLElement {
      * @type {backFunction | null}
      */
     #back_function;
-    /**
-     * @type {CertificateWithCantons | null}
-     */
-    #certificate = null;
-    /**
-     * @type {CantonWithPlaces | null}
-     */
-    #certificate_canton = null;
-    /**
-     * @type {CountryWithCantons | null}
-     */
-    #certificate_country = null;
-    /**
-     * @type {CertificateTypeWithCertificates | null}
-     */
-    #certificate_type = null;
     /**
      * @type {chosenUniversityEntranceQualificationFunction}
      */
@@ -61,9 +41,9 @@ export class UniversityEntranceQualificationElement extends HTMLElement {
      */
     #localization_api;
     /**
-     * @type {CantonWithSchools | null}
+     * @type {[HTMLSelectElement, UniversityEntranceQualificationData[], UniversityEntranceQualificationSelectOption[]][]}
      */
-    #matura_canton = null;
+    #selects;
     /**
      * @type {ShadowRoot}
      */
@@ -72,10 +52,6 @@ export class UniversityEntranceQualificationElement extends HTMLElement {
      * @type {UniversityEntranceQualification}
      */
     #university_entrance_qualification;
-    /**
-     * @type {SchoolWithCountries | null}
-     */
-    #upper_secondary_school = null;
 
     /**
      * @param {CssApi} css_api
@@ -115,6 +91,7 @@ export class UniversityEntranceQualificationElement extends HTMLElement {
         this.#university_entrance_qualification = university_entrance_qualification;
         this.#chosen_university_entrance_qualification_function = chosen_university_entrance_qualification_function;
         this.#back_function = back_function;
+        this.#selects = [];
 
         this.#shadow = this.attachShadow({ mode: "closed" });
         this.#css_api.importCssToRoot(
@@ -134,16 +111,22 @@ export class UniversityEntranceQualificationElement extends HTMLElement {
         }
 
         const post_result = await this.#chosen_university_entrance_qualification_function(
-            {
-                "certificate-type": this.#form_element.inputs["certificate-type"].value,
-                "issue-date": this.#form_element.inputs["issue-date"].valueAsNumber,
-                certificate: this.#form_element.inputs.certificate.value,
-                "matura-canton": this.#form_element.inputs["matura-canton"].value,
-                "upper-secondary-school": this.#form_element.inputs["upper-secondary-school"].value,
-                "certificate-country": this.#form_element.inputs["certificate-country"].value,
-                "certificate-canton": this.#form_element.inputs["certificate-canton"].value,
-                "certificate-place": this.#form_element.inputs["certificate-place"].value
-            }
+            Object.fromEntries([
+                UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_CERTIFICATE,
+                UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_CERTIFICATE_CANTON,
+                UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_CERTIFICATE_COUNTRY,
+                UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_CERTIFICATE_PLACE,
+                UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_CERTIFICATE_TYPE,
+                UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_ISSUE_YEAR,
+                UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_MATURA_CANTON,
+                UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_UPPER_SECONDARY_SCHOOL
+            ].map(select_type => [
+                select_type,
+                this.#form_element.inputs[select_type]?.value ?? null
+            ]).filter(([
+                ,
+                value
+            ]) => value !== null))
         );
 
         if (post_result.ok) {
@@ -195,112 +178,9 @@ export class UniversityEntranceQualificationElement extends HTMLElement {
             )
         );
 
-        const certificate_type_element = this.#form_element.addInput(
-            await this.#localization_api.translate(
-                "Type of certificate"
-            ),
-            "select",
-            "certificate-type"
+        await this.#nextSelect(
+            this.#university_entrance_qualification["select-index"]
         );
-        certificate_type_element.required = true;
-
-        for (const certificate_type of this.#university_entrance_qualification["certificate-types"]) {
-            const option_element = document.createElement("option");
-            option_element.text = await this.#label_service.getCertificateTypeLabel(
-                certificate_type
-            );
-            option_element.value = certificate_type.id;
-            certificate_type_element.appendChild(option_element);
-        }
-
-        certificate_type_element.addEventListener("input", () => {
-            this.#renderCertificates();
-        });
-
-        const issue_date_element = this.#form_element.addInput(
-            await this.#localization_api.translate(
-                "Date of issue (Year)"
-            ),
-            "number",
-            "issue-date"
-        );
-        issue_date_element.inputMode = "numeric";
-        issue_date_element.required = true;
-        issue_date_element.step = 1;
-
-        const certificate_element = this.#form_element.addInput(
-            await this.#localization_api.translate(
-                "Certificate"
-            ),
-            "select",
-            "certificate"
-        );
-        certificate_element.required = true;
-
-        certificate_element.addEventListener("input", () => {
-            this.#renderMaturaCantons();
-        });
-
-        const matura_canton_element = this.#form_element.addInput(
-            await this.#localization_api.translate(
-                "The canton, in which the school is located where you were awarded your matura"
-            ),
-            "select",
-            "matura-canton"
-        );
-        matura_canton_element.required = true;
-
-        matura_canton_element.addEventListener("input", () => {
-            this.#renderUpperSecondarySchool();
-        });
-
-        const upper_secondary_school_element = this.#form_element.addInput(
-            await this.#localization_api.translate(
-                "Upper secondary school"
-            ),
-            "select",
-            "upper-secondary-school"
-        );
-        upper_secondary_school_element.required = true;
-
-        upper_secondary_school_element.addEventListener("input", () => {
-            this.#renderCertificateCountries();
-        });
-
-        const certificate_country_element = this.#form_element.addInput(
-            await this.#localization_api.translate(
-                "The country where you were registered at the time you were awarded your certificate"
-            ),
-            "select",
-            "certificate-country"
-        );
-        certificate_country_element.required = true;
-
-        certificate_country_element.addEventListener("input", () => {
-            this.#renderCertificateCantons();
-        });
-
-        const certificate_canton_element = this.#form_element.addInput(
-            await this.#localization_api.translate(
-                "The canton of your political commune where you were registered at the time you were awarded your certificate"
-            ),
-            "select",
-            "certificate-canton"
-        );
-        certificate_canton_element.required = true;
-
-        certificate_canton_element.addEventListener("input", () => {
-            this.#renderCertificatePlaces();
-        });
-
-        const certificate_place_element = this.#form_element.addInput(
-            await this.#localization_api.translate(
-                "Legal place of residence when the certificate was awarded"
-            ),
-            "select",
-            "certificate-place"
-        );
-        certificate_place_element.required = true;
 
         await this.#form_element.addButtons(
             () => {
@@ -317,194 +197,176 @@ export class UniversityEntranceQualificationElement extends HTMLElement {
         ));
 
         if (this.#university_entrance_qualification.values !== null) {
-            certificate_type_element.value = this.#university_entrance_qualification.values["certificate-type"];
-            await this.#renderCertificates();
-
-            issue_date_element.valueAsNumber = this.#university_entrance_qualification.values["issue-date"];
-
-            certificate_element.value = this.#university_entrance_qualification.values.certificate;
-            await this.#renderMaturaCantons();
-
-            matura_canton_element.value = this.#university_entrance_qualification.values["matura-canton"];
-            await this.#renderUpperSecondarySchool();
-
-            upper_secondary_school_element.value = this.#university_entrance_qualification.values["upper-secondary-school"];
-            await this.#renderCertificateCountries();
-
-            certificate_country_element.value = this.#university_entrance_qualification.values["certificate-country"];
-            await this.#renderCertificateCantons();
-
-            certificate_canton_element.value = this.#university_entrance_qualification.values["certificate-canton"];
-            await this.#renderCertificatePlaces();
-
-            certificate_place_element.value = this.#university_entrance_qualification.values["certificate-place"];
-        } else {
-            await this.#renderCertificates();
+            await this.#initValues();
         }
     }
 
     /**
      * @returns {Promise<void>}
      */
-    async #renderCertificateCantons() {
-        this.#form_element.clearSelectOptions(
-            this.#form_element.inputs["certificate-canton"]
-        );
+    async #initValues() {
+        const [
+            select_element
+        ] = this.#selects[this.#selects.length - 1];
 
-        this.#certificate_country = this.#upper_secondary_school?.countries?.find(country => country.id === this.#form_element.inputs["certificate-country"].value) ?? null;
-
-        await this.#renderCertificatePlaces();
-
-        if (this.#certificate_country === null) {
+        if (select_element.value !== "") {
             return;
         }
 
-        for (const canton of this.#certificate_country.cantons) {
-            const option_element = document.createElement("option");
-            option_element.text = await this.#label_service.getCantonLabel(
-                canton
-            );
-            option_element.value = canton.id;
-            this.#form_element.inputs["certificate-canton"].appendChild(option_element);
+        const value = this.#university_entrance_qualification.values[select_element.name] ?? null;
+
+        if (value === null) {
+            return;
         }
+
+        select_element.value = value;
+
+        await this.#selected(
+            select_element
+        );
+
+        await this.#initValues();
     }
 
     /**
+     * @param {number} select_index
      * @returns {Promise<void>}
      */
-    async #renderCertificateCountries() {
-        this.#form_element.clearSelectOptions(
-            this.#form_element.inputs["certificate-country"]
+    async #nextSelect(select_index) {
+        const [
+            select_to_data_index,
+            select_options
+        ] = this.#university_entrance_qualification.selects[select_index];
+
+        const [
+            select_type,
+            data_index
+        ] = this.#university_entrance_qualification["select-to-data"][select_to_data_index];
+
+        const data = this.#university_entrance_qualification.data[data_index];
+
+        let label, get_option_label;
+        switch (select_type) {
+            case UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_CERTIFICATE:
+                label = "Certificate";
+                get_option_label = "getCertificateLabel";
+                break;
+
+            case UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_CERTIFICATE_CANTON:
+                label = "The canton of your political commune where you were registered at the time you were awarded your certificate";
+                get_option_label = "getCantonLabel";
+                break;
+
+            case UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_CERTIFICATE_COUNTRY:
+                label = "The country where you were registered at the time you were awarded your certificate";
+                get_option_label = "getCountryLabel";
+                break;
+
+            case UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_CERTIFICATE_PLACE:
+                label = "Legal place of residence when the certificate was awarded";
+                get_option_label = "getPlaceLabel";
+                break;
+
+            case UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_CERTIFICATE_TYPE:
+                label = "Type of certificate";
+                get_option_label = "getCertificateTypeLabel";
+                break;
+
+            case UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_ISSUE_YEAR:
+                label = "Year of issue";
+                get_option_label = "getIssueYearLabel";
+                break;
+
+            case UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_MATURA_CANTON:
+                label = "The canton, in which the school is located where you were awarded your matura";
+                get_option_label = "getCantonLabel";
+                break;
+
+            case UNIVERSITY_ENTRANCE_QUALIFICATION_SELECT_TYPE_UPPER_SECONDARY_SCHOOL:
+                label = "Upper secondary school";
+                get_option_label = "getSchoolLabel";
+                break;
+
+            default:
+                return;
+        }
+
+        const select_element = this.#form_element.addInput(
+            await this.#localization_api.translate(
+                label
+            ),
+            "select",
+            select_type
         );
+        select_element.required = true;
 
-        this.#upper_secondary_school = this.#matura_canton?.schools?.find(school => school.id === this.#form_element.inputs["upper-secondary-school"].value) ?? null;
+        for (const select_option of select_options) {
+            const option = data[typeof select_option === "number" ? select_option : select_option[0]];
 
-        await this.#renderCertificateCantons();
-
-        if (this.#upper_secondary_school === null) {
-            return;
-        }
-
-        for (const country of this.#upper_secondary_school.countries) {
             const option_element = document.createElement("option");
-            option_element.text = await this.#label_service.getCountryLabel(
-                country
+            option_element.text = await this.#label_service[get_option_label](
+                option
             );
-            option_element.value = country.id;
-            this.#form_element.inputs["certificate-country"].appendChild(option_element);
+            option_element.value = option.id;
+            select_element.appendChild(option_element);
         }
+
+        select_element.addEventListener("input", () => {
+            this.#selected(
+                select_element
+            );
+        });
+
+        this.#selects.push([
+            select_element,
+            data,
+            select_options
+        ]);
     }
 
     /**
+     * @param {HTMLSelectElement} select_element
      * @returns {Promise<void>}
      */
-    async #renderCertificatePlaces() {
-        this.#form_element.clearSelectOptions(
-            this.#form_element.inputs["certificate-place"]
-        );
+    async #selected(select_element) {
+        const select_index = this.#selects.findIndex(([
+            _select_element
+        ]) => _select_element === select_element);
 
-        this.#certificate_canton = this.#certificate_country?.cantons?.find(canton => canton.id === this.#form_element.inputs["certificate-canton"].value) ?? null;
-
-        if (this.#certificate_canton === null) {
+        if (select_index === -1) {
             return;
         }
 
-        for (const place of this.#certificate_canton.places) {
-            const option_element = document.createElement("option");
-            option_element.text = await this.#label_service.getPlaceLabel(
-                place
-            );
-            option_element.value = place.id;
-            this.#form_element.inputs["certificate-place"].appendChild(option_element);
+        if (this.#selects.length > (select_index + 1)) {
+            for (const [
+                _select_element
+            ] of this.#selects.slice(select_index + 1)) {
+                _select_element.parentElement.remove();
+            }
+
+            this.#selects = this.#selects.slice(0, select_index + 1);
         }
-    }
 
-    /**
-     * @returns {Promise<void>}
-     */
-    async #renderCertificates() {
-        this.#form_element.inputs["issue-date"].max = 0;
-        this.#form_element.inputs["issue-date"].min = 0;
-        this.#form_element.inputs["issue-date"].value = "";
+        const [
+            ,
+            data,
+            select_options
+        ] = this.#selects[select_index];
 
-        this.#form_element.clearSelectOptions(
-            this.#form_element.inputs.certificate
-        );
+        const data_index = data.findIndex(option => option.id === select_element.value);
 
-        this.#certificate_type = this.#university_entrance_qualification["certificate-types"].find(certificate_type => certificate_type.id === this.#form_element.inputs["certificate-type"].value) ?? null;
+        const select_option = select_options.find(_select_option => (typeof _select_option === "number" ? _select_option : _select_option[0]) === data_index) ?? null;
 
-        await this.#renderMaturaCantons();
-
-        if (this.#certificate_type === null) {
+        if (select_option === null || typeof select_option === "number") {
             return;
         }
 
-        this.#form_element.inputs["issue-date"].max = this.#certificate_type["max-issue-date"];
-        this.#form_element.inputs["issue-date"].min = this.#certificate_type["min-issue-date"];
-
-        for (const certificate of this.#certificate_type.certificates) {
-            const option_element = document.createElement("option");
-            option_element.text = await this.#label_service.getCertificateLabel(
-                certificate
-            );
-            option_element.value = certificate.id;
-            this.#form_element.inputs.certificate.appendChild(option_element);
-        }
-    }
-
-    /**
-     * @returns {Promise<void>}
-     */
-    async #renderMaturaCantons() {
-        this.#form_element.clearSelectOptions(
-            this.#form_element.inputs["matura-canton"]
+        await this.#nextSelect(
+            select_option[1]
         );
-
-        this.#certificate = this.#certificate_type?.certificates?.find(certificate => certificate.id === this.#form_element.inputs.certificate.value) ?? null;
-
-        await this.#renderUpperSecondarySchool();
-
-        if (this.#certificate === null) {
-            return;
-        }
-
-        for (const canton of this.#certificate.cantons) {
-            const option_element = document.createElement("option");
-            option_element.text = await this.#label_service.getCantonLabel(
-                canton
-            );
-            option_element.value = canton.id;
-            this.#form_element.inputs["matura-canton"].appendChild(option_element);
-        }
-    }
-
-    /**
-     * @returns {Promise<void>}
-     */
-    async #renderUpperSecondarySchool() {
-        this.#form_element.clearSelectOptions(
-            this.#form_element.inputs["upper-secondary-school"]
-        );
-
-        this.#matura_canton = this.#certificate?.cantons?.find(canton => canton.id === this.#form_element.inputs["matura-canton"].value) ?? null;
-
-        await this.#renderCertificateCountries();
-
-        if (this.#matura_canton === null) {
-            return;
-        }
-
-        for (const school of this.#matura_canton.schools) {
-            const option_element = document.createElement("option");
-            option_element.text = await this.#label_service.getSchoolLabel(
-                school
-            );
-            option_element.value = school.id;
-            this.#form_element.inputs["upper-secondary-school"].appendChild(option_element);
-        }
     }
 }
 
-export const UNIVERSITY_ENTRANCE_QUALIFICATION_ELEMENT_TAG_NAME = `${ELEMENT_TAG_NAME_PREFIX}${PAGE_UNIVERSITY_ENTRANCE_QUALIFICATION}`;
+export const UNIVERSITY_ENTRANCE_QUALIFICATION_ELEMENT_TAG_NAME = `flux-studis-selfservice-${PAGE_UNIVERSITY_ENTRANCE_QUALIFICATION}`;
 
 customElements.define(UNIVERSITY_ENTRANCE_QUALIFICATION_ELEMENT_TAG_NAME, UniversityEntranceQualificationElement);