+++ src/Adapter/UniversityEntranceQualification/UniversityEntranceQualificationSelectOption.mjs
@@ -1,0 +1,3 @@
+/**
+ * @typedef {[number, number] | number} UniversityEntranceQualificationSelectOption
+ */